Move `handle_rename_project_entry` to LspStore
